requirements.txt
quart>=0.10.0
botbuilder-dialogs>=4.11.0
botbuilder-ai>=4.11.0
botbuilder-core>=4.11.0
botbuilder-dialogs>=4.11.0
botbuilder-applicationinsights>=4.11.0
botbuilder-integration-applicationinsights-aiohttp>=4.11.0
datatypes-date-time>=1.0.0.a2
aiohttp>=3.6.2
nest-asyncio>=1.4.1
pytest>=5.4.3
